+++ scripts/main.js
@@ -115,7 +115,7 @@ function createWindow() {
 		{
 			appClosing = true;
 
-			win.webContents.executeJavaScript('var saved = reading.saveReadingProgress(); settings.purgeTemporaryFiles(); cache.purge(); ebook.closeAllRenders(); workers.closeAllWorkers(); saved;', false).then(function(value) {
+			win.webContents.executeJavaScript('var saved = reading.progress.save(); settings.purgeTemporaryFiles(); cache.purge(); ebook.closeAllRenders(); workers.closeAllWorkers(); saved;', false).then(function(value) {
 
 				if(!value)
 					app.quit();